Nguyên nạp tiền

+++ app/src/main/java/com/project/veggieempiregr5/BankUtil.java
@@ -1,6 +1,6 @@
 package com.project.veggieempiregr5;
 
 public class BankUtil {
-    public static Double soTienNap =0.0;
+    public static Integer soTienNap =0;
     public static String tenNH ="";
 }
